% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{get_ellipses}
\alias{get_ellipses}
\title{Obtain dataframe of kernel ellipses centered at the DPC gridpoints}
\usage{
get_ellipses(dpc_grid, fit, eval)
}
\arguments{
\item{dpc_grid}{Discrete Process Convolution grid}

\item{fit}{Fitted model}

\item{eval}{Size of ellipses: k[s_i, s_j, \phi] = eval}
}
\value{
Dataframe with coordinates of points that are points in the ellipses; gp identifies the gridpoint
}
\description{
Obtain dataframe of kernel ellipses centered at the DPC gridpoints
}
