% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{get_ellipses_plot}
\alias{get_ellipses_plot}
\title{Plot kernel ellipses centered at the DPC gridpoints}
\usage{
get_ellipses_plot(dpc_grid, fit, eval = 0.7)
}
\arguments{
\item{dpc_grid}{Discrete Process Convolution grid}

\item{fit}{Fitted model}

\item{eval}{Size of ellipses: k[s_i, s_j, \phi] = eval}
}
\value{
Plot of ellipses
}
\description{
Plot kernel ellipses centered at the DPC gridpoints
}
\examples{
s = expand.grid(lat = -1:1, lon=-1:1)
y = c(1,0,0,0,1,0,0,0,1)
dpc_grid = get_grid(c(-1,1), c(-1,1), spacing = 2)
priors = get_priors(dpc_grid)
iso_kernel_matrix = get_kernel_matrix(s, dpc_grid)
fit = get_mcmc(s, dpc_grid, y, 10, 1000, priors, 100, 1)
get_interpolation_plot(s, dpc_grid, fit)
get_ellipses_plot(dpc_grid, fit)
}
