% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{get_em}
\alias{get_em}
\title{Optimize kernel parameters}
\usage{
get_em(nburn, priors, s, y, iso_kernel_matrix)
}
\arguments{
\item{nburn}{Number of iterations}

\item{priors}{Prior distributions for gamma (Gaussian) and tau (Gamma)}

\item{s}{Coordinates of observations}

\item{y}{Values of observations}

\item{iso_kernel_matrix}{Isotropic kernel matrix}
}
\value{
List with optimized kernel parameters, as well as corresponding mean values of gamma and tau
}
\description{
Optimize kernel parameters
}
\examples{
s = expand.grid(lat = c(-1:1), lon = c(-1:1))
y = c(1,0,0,0,1,0,0,0,1)
dpc_grid = get_grid(c(-1,1), c(-1,1), spacing = 2)
priors = get_priors(dpc_grid)
iso_kernel_matrix = get_kernel_matrix(s, dpc_grid)
fit = get_em(10, priors, s, y, iso_kernel_matrix)
cbind(get_estimates(s, dpc_grid, fit), obs = y)

}
