% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{get_estimates}
\alias{get_estimates}
\title{Get posterior predictive mean estimates at location(s) in the domain}
\usage{
get_estimates(s, dpc_grid, fit)
}
\arguments{
\item{s}{Locations}

\item{dpc_grid}{Discrete Process Convolution grid}

\item{fit}{Fitted model}
}
\value{
Data frame with coordinates and posterior predictive means
}
\description{
Get posterior predictive mean estimates at location(s) in the domain
}
\examples{
s = data.frame(lat = rep(0, 3), lon = c(-1:1))
y = c(1,1,0)
dpc_grid = get_grid(c(-1,1), c(0,0), spacing = 2)
priors = get_priors(dpc_grid)
iso_kernel_matrix = get_kernel_matrix(s, dpc_grid)
fit = get_mcmc(s, dpc_grid, y, 10, 1000, priors, 100, 1)
cbind(get_estimates(s, dpc_grid, fit), obs = y)

}
