% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammatau.R
\name{get_gamma_sample}
\alias{get_gamma_sample}
\title{Sample gamma from its full conditional distribution}
\usage{
get_gamma_sample(prior_mean, prior_precision, kernel_matrix, tau, y,
  return_mean = FALSE)
}
\arguments{
\item{prior_mean}{Prior mean of gamma}

\item{prior_precision}{Prior precision of gamma}

\item{kernel_matrix}{DPC kernel matrix}

\item{tau}{Sampled value of tau}

\item{y}{Observation value(s)}

\item{return_mean}{Provide full conditional mean as pseudo-sample?}
}
\value{
Either the full conditional mean of gamma or a sample of it
}
\description{
Sample gamma from its full conditional distribution
}
\examples{
km = list(mat = Matrix::Diagonal(2, 1))
prior_pr = Matrix::Diagonal(2, 0.3)
tau = 5
post_pr = get_posterior_precision(prior_pr, km, tau)
get_gamma_sample(prior_mean = rep(0, 2), prior_precision = prior_pr, 
                 kernel_matrix = km, tau = tau, y = c(0.9, 0.7), return_mean = TRUE)

}
