% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{get_grid}
\alias{get_grid}
\title{Produce a regular Discrete Process Convolution grid.}
\usage{
get_grid(longitude_bounds, latitude_bounds, spacing)
}
\arguments{
\item{longitude_bounds}{Minimum and maximum longitude}

\item{latitude_bounds}{Minimum and maximum latitude}

\item{spacing}{Grid resolution, i.e., distance between adjacent gridpoints}
}
\value{
List with gridded coodinates, as well as metadata
}
\description{
Produce a regular Discrete Process Convolution grid.
}
\examples{
get_grid(c(-1,1), c(2,4), 1)
}
