% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{get_influence_plot}
\alias{get_influence_plot}
\title{Plot the influence of a gridded Gaussian variable on the interpolation}
\usage{
get_influence_plot(dpc_grid, lat, lon, fit = NULL)
}
\arguments{
\item{dpc_grid}{Discrete Process Convolution grid}

\item{lat}{Latitute of desired gridpoint}

\item{lon}{Longitude of desired gridpoint}

\item{fit}{(Optional) fitted model}
}
\value{
A ggplot showing the weight of the selected variable on the fit
}
\description{
Plot the influence of a gridded Gaussian variable on the interpolation
}
\examples{
get_influence_plot(get_grid(c(-1,1), c(-1,1), spacing = 2), -1, -1, fit = NULL)
}
