% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{get_interpolation_plot}
\alias{get_interpolation_plot}
\title{Plot the interpolated surface}
\usage{
get_interpolation_plot(obs_coord, dpc_grid, fit, fineness = 16,
  contour_binwidth = NULL, do_grid = TRUE)
}
\arguments{
\item{obs_coord}{Coordinates of original observations}

\item{dpc_grid}{Discrete Process Convolution grid}

\item{fit}{Fitted model}

\item{fineness}{Resolution of interpolation (1 = coarse, 10 = fine)}

\item{contour_binwidth}{Width of bins for contour plot}

\item{do_grid}{Plot DPC gridpoints?}
}
\value{
Plot of gridded posterior predictive means
}
\description{
Plot the interpolated surface
}
\examples{
s = expand.grid(lat = -1:1, lon=-1:1)
y = c(1,0,0,0,1,0,0,0,1)
dpc_grid = get_grid(c(-1,1), c(-1,1), spacing = 2)
priors = get_priors(dpc_grid)
iso_kernel_matrix = get_kernel_matrix(s, dpc_grid)
fit = get_mcmc(s, dpc_grid, y, 10, 1000, priors, 100, 1)
get_interpolation_plot(s, dpc_grid, fit, contour_binwidth = 0.1)
}
