% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{get_kernel_evaluation}
\alias{get_kernel_evaluation}
\title{Get kernel evaluation(s) centered at s_i and evaluated at one or many gridpoints s_j}
\usage{
get_kernel_evaluation(s_i, s_j, grid_spacing, phi_i)
}
\arguments{
\item{s_i}{Point where the kernel is centered at}

\item{s_j}{Grid points(s) where the kernel is evaluated at}

\item{grid_spacing}{DPC grid resolution}

\item{phi_i}{Kernel parameters at location s_i}
}
\value{
Kernel evaluation(s) at the gridpoint(s) s_j
}
\description{
Get kernel evaluation(s) centered at s_i and evaluated at one or many gridpoints s_j
}
\examples{
get_kernel_evaluation(s_i = list(lat = 0, lon = 0), s_j = list(lat = 0, lon = 0), 
                      grid_spacing = 1, phi_i = c(0, 0, 0, 0)) == 1
get_kernel_evaluation(s_i = list(lat = 0, lon = 0), s_j = list(lat = 0, lon = 1), 
                      grid_spacing = 1, phi_i = c(0, 0, 0, 0)) == 0
get_kernel_evaluation(s_i = list(lat = 0, lon = 0), s_j = list(lat = 0, lon = 1), 
                      grid_spacing = 2, phi_i = c(0, 0, 0, 0)) == 0.75

}
