% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{get_kernel_matrix}
\alias{get_kernel_matrix}
\title{Construct a sparse kernel matrix}
\usage{
get_kernel_matrix(s, dpc_grid, phi = NULL)
}
\arguments{
\item{s}{Observation locations}

\item{dpc_grid}{Discrete Process Convolution grid}

\item{phi}{Optional kernel parameters}
}
\value{
List with sparse matrix (nrow = number of observations, ncol = number of gridpoints), plus
        metadata
}
\description{
Construct a sparse kernel matrix
}
\examples{
s = data.frame(lon = 1:3, lat = 1:3)
dpc_grid = get_grid(c(1, 3), c(1, 3), 1)
get_kernel_matrix(s, dpc_grid)$mat

}
