% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{get_kernel_vector}
\alias{get_kernel_vector}
\title{Obtain the vector of kernel evaluations for all the gridpoints in the neighborhood of observation s_i}
\usage{
get_kernel_vector(s_i, neighbors_i, dpc_grid, phi_i)
}
\arguments{
\item{s_i}{Observation location}

\item{neighbors_i}{List of gridpoints in the neighborhood of s_i}

\item{dpc_grid}{Discrete Process Convolution grid}

\item{phi_i}{Kernel parameters}
}
\value{
List with normalized vector of kernel evaluations, as well as sum of unnormalized evaluations
}
\description{
Obtain the vector of kernel evaluations for all the gridpoints in the neighborhood of observation s_i
}
\examples{
s_i = list(lat = 0, lon = 0)
dpc_grid = get_grid(c(-1, 1), c(0, 0), 1)
neighbors_i = c(1,2,3)
phi_i = c(0.5, 1, 0, 0)
get_kernel_vector(s_i, neighbors_i, dpc_grid, phi_i)
}
