% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{get_list2mat}
\alias{get_list2mat}
\title{Convert list of observation locations into data frame}
\usage{
get_list2mat(s)
}
\arguments{
\item{s}{List of observation locations}
}
\value{
Data frame of obs locations
}
\description{
Convert list of observation locations into data frame
}
