% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{get_mat2list}
\alias{get_mat2list}
\title{Convert matrix/data.frame observation locations into list}
\usage{
get_mat2list(s)
}
\arguments{
\item{s}{Observation locations}
}
\value{
List of obs locations
}
\description{
Convert matrix/data.frame observation locations into list
}
