% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{get_neighbors_i}
\alias{get_neighbors_i}
\title{Identify the gridpoints in the neighborhood of location s_i}
\usage{
get_neighbors_i(s_i, dpc_grid)
}
\arguments{
\item{s_i}{Observation location}

\item{dpc_grid}{Discrete Process Convolution grid}
}
\value{
Vector of indices that identify the grispoint neighbors of s_i
}
\description{
Identify the gridpoints in the neighborhood of location s_i
}
\examples{
s_i = list(lat = 0, lon = 0)
dpc_grid = get_grid(c(0, 5), c(0, 5), 1)
get_neighbors_i(s_i, dpc_grid)
}
