% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{get_neighbors_of_gridpoints}
\alias{get_neighbors_of_gridpoints}
\title{Identify the observation locations that are in the neighborhood of each gridpoint}
\usage{
get_neighbors_of_gridpoints(neighborsI, ngridpoints)
}
\arguments{
\item{neighborsI}{List of gridpoint neighbors to each observation}

\item{ngridpoints}{Number of gridpoints}
}
\value{
List of observation locations that are neighbors to each gridpoint
}
\description{
Identify the observation locations that are in the neighborhood of each gridpoint
}
\examples{
neighborsI = list(list(1, 2), list(2, 3))
get_neighbors_of_gridpoints(neighborsI, 3)
}
