% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{get_point_estimate}
\alias{get_point_estimate}
\title{Get a point estimate for a single location}
\usage{
get_point_estimate(lon, lat, dpc_grid, fit)
}
\arguments{
\item{lon}{Longitude of point}

\item{lat}{Latitude of point}

\item{dpc_grid}{Discrete Process Convolution grid}

\item{fit}{Fitted model}
}
\value{
Posterior predictive mean at given location
}
\description{
Get a point estimate for a single location
}
\examples{
s = data.frame(lat = rep(0, 3), lon = c(-1:1))
y = c(1,1,0)
dpc_grid = get_grid(c(-1,1), c(0,0), spacing = 2)
priors = get_priors(dpc_grid)
iso_kernel_matrix = get_kernel_matrix(s, dpc_grid)
fit = get_mcmc(s, dpc_grid, y, 10, 1000, priors, 100, 1)
get_point_estimate(lon = s$lon[1], lat = s$lat[1], dpc_grid = dpc_grid, fit = fit)

}
