% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammatau.R
\name{get_posterior_precision}
\alias{get_posterior_precision}
\title{Compute the full conditional (given tau, rho, and data) precision of gamma}
\usage{
get_posterior_precision(prior_precision, kernel_matrix, tau)
}
\arguments{
\item{prior_precision}{Prior precision matrix for gamma}

\item{kernel_matrix}{Discrete Process Convolution kernel matrix (fixed values of rho)}

\item{tau}{Sampled value of tau}
}
\value{
Full conditional precision of gamma
}
\description{
Compute the full conditional (given tau, rho, and data) precision of gamma
}
\examples{
get_posterior_precision(Matrix::Diagonal(2, 3), list(mat = Matrix::Diagonal(2, 1)), 5)

}
