% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammatau.R
\name{get_prior_precision_matrix}
\alias{get_prior_precision_matrix}
\title{Construct gamma's precision matrix for a given DPC grid}
\usage{
get_prior_precision_matrix(dpc_grid, diag_value, nearest_neighbor_factor)
}
\arguments{
\item{dpc_grid}{Discrete Process Convolution grid}

\item{diag_value}{Diagonal value in precision matrix}

\item{nearest_neighbor_factor}{Off-diagonal factor for adjacent gridpoints}
}
\value{
A sparse precision matrix
}
\description{
Construct gamma's precision matrix for a given DPC grid
}
\examples{
dpc_grid = get_grid(c(0,1), c(1,2), 1)
get_prior_precision_matrix(dpc_grid, 3, 0.2)

}
