% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammatau.R
\name{get_priors}
\alias{get_priors}
\title{Build basic priors for gamma and tau}
\usage{
get_priors(dpc_grid, precision_diagonal_value = 1e-06,
  precision_nearest_neighbor_factor = 0)
}
\arguments{
\item{dpc_grid}{Discrete Process Convolution grid}

\item{precision_diagonal_value}{Diagonal value for gamma's prior precision matrix}

\item{precision_nearest_neighbor_factor}{Off-diagonal factor for gamma's prior precision matrix, in (0,1)}
}
\value{
List with prior mean and precision for gamma, and prior shape and rate parameters for tau
}
\description{
Build basic priors for gamma and tau
}
\examples{
dpc_grid = get_grid(c(0,1), c(1,2), 1)
priors = get_priors(dpc_grid, 5, 0.4)
priors$gamma$precision
(prior_variance = solve(priors$gamma$precision))

}
