% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{get_Sigma_inverse}
\alias{get_Sigma_inverse}
\title{Lemos and Sanso' (2012) Inverse Sigma matrix}
\usage{
get_Sigma_inverse(phi_i, grid_spacing)
}
\arguments{
\item{phi_i}{Kernel parameters}

\item{grid_spacing}{Distance between adjacent grid points (a.k.a. resolution)}
}
\value{
A 2x2 matrix used to compute the distance between two points
}
\description{
Lemos and Sanso' (2012) Inverse Sigma matrix
}
