% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammatau.R
\name{get_tau_sample}
\alias{get_tau_sample}
\title{Sample tau from its full conditional distribution}
\usage{
get_tau_sample(kernel_matrix, gamma, y, prior_tau_a, prior_tau_b,
  return_mean = FALSE)
}
\arguments{
\item{kernel_matrix}{DPC kernel matrix}

\item{gamma}{Sample of gamma}

\item{y}{Observation values}

\item{prior_tau_a}{Prior shape parameter}

\item{prior_tau_b}{Prior rate parameter}

\item{return_mean}{Provide full conditional mean as pseudo-sample?}
}
\value{
Either the full conditional mean of tau or a sample of it
}
\description{
Sample tau from its full conditional distribution
}
\examples{
km = list(mat = Matrix::Diagonal(2, 1))
get_tau_sample(kernel_matrix = km, gamma = c(0.849, 0.66), y = c(0.9, 0.7), 
               prior_tau_a = 2, prior_tau_b = 2, return_mean = TRUE)
                 
}
