% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{haversine}
\alias{haversine}
\title{Haversine distance between two points on the Earth's surface}
\usage{
haversine(s_i, s_j)
}
\arguments{
\item{s_i}{Point i, identified by latitude and longitude}

\item{s_j}{Point j}
}
\value{
Distance between points i and j, in km
}
\description{
Haversine distance between two points on the Earth's surface
}
\examples{
d1 = haversine(s_i = list(lat = 0, lon = 10), s_j = list(lat = 0, lon = 11))
d2 = haversine(s_i = list(lat = 10, lon = 10), s_j = list(lat = 11, lon = 10))
abs(2 * pi * 6371 / 360 - d1) < 0.001
abs(2 * pi * 6371 / 360 - d2) < 0.001

}
